model_deployment/requirements.txt
streamlit
pandas
seaborn
matplotlib
numpy
plotly
Pillow
scikit-learn
tensorflow